Create ETL_toll_data.py
Project that aims to de-congest the national highways by analyzing the road traffic data from different toll plazas. Each highway is operated by a different toll operator with a different IT setup that uses different file formats. Task is to collect data available in different formats and consolidate it into a single file.